Implemented Logger Component
