+++ Basestation_Software.Web/Core/Components/LoggerTester.razor
@@ -1,5 +1,4 @@
-﻿@inject LoggerService _LoggerService
-<div class="card full-height">
+﻿<div class="card full-height">
 	<div class="card-header">
 		<h3>Logger Tester</h3>
 	</div>
@@ -11,11 +10,11 @@
 @code {
 	protected override async Task OnInitializedAsync()
 	{
-		_LoggerService.Log("Tester initialized!");
+		LoggerCore.LoggerService.Log("Tester initialized!");
 	}
 
 	public void AddExampleLog()
 	{
-		_LoggerService.Log("Example message");
+		LoggerCore.LoggerService.Log("Example message");
 	}
 }
