+++ Basestation_Software.Web/Program.cs
@@ -39,7 +39,8 @@ builder.Services.AddSingleton<PingService>();
 builder.Services.AddSingleton<OpService>();
 builder.Services.AddRoveComm();
 builder.Services.AddBlazoredToast();
-builder.Services.AddSingleton<LoggerService>();
+//builder.Services.AddSingleton<LoggerService_Deprecated>();
+//builder.Services.AddSingleton<LoggerCore.LoggerService>();
 
 var app = builder.Build();
